refactor + goimports + gofmt
