+++ mediator/mediator.go
@@ -44,8 +44,8 @@ type Mediator struct {
 }
 
 type Info struct {
-	Scribes map[string]string
-	ScribesCounter map[string]int64
+	Scribes              map[string]string
+	ScribesCounter       map[string]int64
 	ScribeResponsibility map[string]string
 }
 
@@ -135,7 +135,7 @@ func (m *Mediator) getConnection(s string) (*grpc.ClientConn, error) {
 }
 
 func (m *Mediator) startPingingSubcribers() {
-        for range time.Tick(5 * time.Second) {
+	for range time.Tick(5 * time.Second) {
 		m.mux.Lock()
 		m.pingSubscribers()
 		m.mux.Unlock()
